refactor: 将build_collate_fn挪到dataloader.py中 refactor: 删除了一些没用的代码 refactor: MMT开发暂缓，挪到Outdated中
