--- train.sh
@@ -1,1 +1,0 @@
-CUDA_VISIBLE_DEVICES=0,1 python -m torch.distributed.launch --nproc_per_node 2 train_mmt.py
