+++ utils.py
@@ -131,65 +131,3 @@ def show_input_shape(**kwargs):
             print("")
     print("***************************************\n")
 
-
-def build_collate_fn(pad_id: int, include_id: bool):
-    def func1(data):
-        batch_size = len(data)
-        # video id
-        id_data = [i[2] for i in data]
-
-        # video feature
-        feat_dim = data[0][0].shape[1]
-        feat_data = [i[0] for i in data]
-        feat_len = [len(i) for i in feat_data]
-        max_len = max(feat_len)
-        feat_ts = torch.zeros([batch_size, max_len, feat_dim], dtype=torch.float)
-        feat_mask_ts = torch.ones([batch_size, max_len], dtype=torch.long)
-        for i in range(batch_size):
-            feat_ts[i, :feat_len[i]] = feat_data[i]
-            feat_mask_ts[i, :feat_len[i]] = 0
-        feat_mask_ts = (feat_mask_ts == 1)
-
-        # text
-        text_data = [i[1] for i in data]
-        text_len = [len(i) for i in text_data]
-        max_len = max(text_len)
-        text_ts = torch.ones([batch_size, max_len], dtype=torch.long) * pad_id
-        for i in range(batch_size):
-            text_ts[i, :text_len[i]] = text_data[i]
-        text_mask_ts = (text_ts == pad_id)
-        return feat_ts, text_ts, feat_mask_ts, text_mask_ts, id_data
-
-    def func2(data):
-        """
-        :param data:
-        :return tuple(N T E, N T):
-        """
-        batch_size = len(data)
-
-        # video feature
-        feat_dim = data[0][0].shape[1]
-        feat_data = [i[0] for i in data]
-        feat_len = [len(i) for i in feat_data]
-        max_len = max(feat_len)
-        feat_ts = torch.zeros([batch_size, max_len, feat_dim], dtype=torch.float)
-        feat_mask_ts = torch.ones([batch_size, max_len], dtype=torch.long)
-        for i in range(batch_size):
-            feat_ts[i, :feat_len[i]] = feat_data[i]
-            feat_mask_ts[i, :feat_len[i]] = 0
-        feat_mask_ts = (feat_mask_ts == 1)
-
-        # text
-        text_data = [i[1] for i in data]
-        text_len = [len(i) for i in text_data]
-        max_len = max(text_len)
-        text_ts = torch.ones([batch_size, max_len], dtype=torch.long) * pad_id
-        for i in range(batch_size):
-            text_ts[i, :text_len[i]] = text_data[i]
-        text_mask_ts = (text_ts == pad_id)
-        return feat_ts, text_ts, feat_mask_ts, text_mask_ts
-
-    if include_id is True:
-        return func1
-    else:
-        return func2
